Timestamp operations and simultaneous keyframe sorting

+++ StoryboardSystem/API/IStoryboardParams.cs
@@ -3,7 +3,7 @@
 namespace StoryboardSystem; 
 
 public interface IStoryboardParams {
-    float Convert(int measures, int beats, float ticks, float seconds);
+    float Convert(float measures, float beats, float ticks, float seconds);
 
     Object GetExternalObject(string name);
 }
